refactor state module out
