+++ client/.changeset/config.json
@@ -16,7 +16,6 @@
     "@wikijump/css",
     "@wikijump/prism",
     "@wikijump/sheaf",
-    "@wikijump/state",
     "@wikijump/docs",
     "dev-sandbox"
   ]
