+++ client/modules/api/package.json
@@ -17,5 +17,9 @@
   "devDependencies": {
     "@mockoon/cli": "^1.1.0",
     "swagger-typescript-api": "^9.2.0"
+  },
+  "dependencies": {
+    "@wikijump/util": "workspace:*",
+    "svelte-i18n": "^3.3.12"
   }
 }
