+++ client/modules/api/src/index.ts
@@ -83,3 +83,4 @@ export type {
 } from "../vendor/api"
 export * from "./api"
 export * from "./asset"
+export * from "./i18n"
