+++ client/modules/cm-lang-ftml/package.json
@@ -19,7 +19,7 @@
     "@wikijump/components": "workspace:*",
     "@wikijump/ftml-wasm-worker": "workspace:*",
     "@wikijump/prism": "workspace:*",
-    "@wikijump/state": "workspace:*",
+    "@wikijump/api": "workspace:*",
     "@wikijump/threads-worker-module": "workspace:*",
     "@wikijump/util": "workspace:*",
     "cm-tarnation": "^0.1.0"
