+++ client/modules/components/src/index.ts
@@ -15,3 +15,5 @@ export * from "./lib/focus"
 export * from "./lib/popper"
 export * from "./lib/tippy"
 export * from "./lib/portal"
+export * from "./lib/media"
+export * from "./lib/useragent"
