+++ client/modules/sheaf/src/components/Sheaf.svelte
@@ -7,7 +7,8 @@
   import type { SheafBindings } from "../extensions/bindings"
   import { setContext } from "svelte"
   import type { Readable } from "svelte/store"
-  import { matchBreakpoint, PreferenceHandler } from "@wikijump/state"
+  import { PreferenceHandler } from "@wikijump/util"
+  import { matchBreakpoint } from "@wikijump/components"
   import type { SheafContext } from "../context"
   import { getDefaultSheafSettings } from "../context"
   import PaneEditor from "./PaneEditor.svelte"
