--- client/modules/state/README.md
@@ -1,3 +1,0 @@
-# @wikijump/state
-
-Client state helpers and handlers, for Wikijump.
