--- client/modules/state/package.json
@@ -1,21 +1,0 @@
-{
-  "name": "@wikijump/state",
-  "license": "agpl-3.0-or-later",
-  "description": "State handlers for Wikijump.",
-  "version": "0.0.0",
-  "keywords": [
-    "wikijump"
-  ],
-  "private": true,
-  "scripts": {},
-  "type": "module",
-  "main": "src/index.ts",
-  "devDependencies": {
-    "svelte": "^3.43.1"
-  },
-  "dependencies": {
-    "is-what": "^4.0.0",
-    "svelte-i18n": "^3.3.12",
-    "@wikijump/util": "workspace:*"
-  }
-}
