--- client/modules/state/src/index.ts
@@ -1,4 +1,0 @@
-export * from "./i18n"
-export * from "./media"
-export * from "./pref"
-export * from "./useragent"
