--- client/modules/state/tsconfig.json
@@ -1,7 +1,0 @@
-{
-  "compilerOptions": {
-    "composite": true
-  },
-  "extends": "../../tsconfig.json",
-  "include": ["src/**/*", "tests/**/*", "../../misc/types/**/*"]
-}
