+++ client/modules/util/src/index.ts
@@ -1,6 +1,7 @@
 export * from "./decorators"
 export * from "./html"
 export * from "./timeout"
+export * from "./pref"
 
 // https://gist.github.com/hyamamoto/fd435505d29ebfa3d9716fd2be8d42f0#gistcomment-2694461
 /** Very quickly generates a (non-secure) hash from the given string. */
